+++ src/contrib/google/sheets/ui/AppendSpreadsheetOptions.test.tsx
@@ -28,7 +28,7 @@ import {
 import { getToggleOptions } from "@/components/fields/schemaFields/getToggleOptions";
 import { dereference } from "@/validators/generic";
 import { BASE_SHEET_SCHEMA } from "@/contrib/google/sheets/core/schemas";
-import SheetsFileWidget from "@/contrib/google/sheets/ui/SheetsFileWidget";
+import SheetsFileWidget from "@/contrib/google/sheets/ui/SpreadsheetPickerWidget";
 import { render } from "@/pageEditor/testHelpers";
 import { validateRegistryId } from "@/types/helpers";
 import { services, sheets } from "@/background/messenger/api";
