+++ src/contrib/google/sheets/ui/SheetServiceOptions.tsx
@@ -17,7 +17,7 @@
 
 import React, { useState } from "react";
 import { type BlockOptionProps } from "@/components/fields/schemaFields/genericOptionsFactory";
-import SheetsFileWidget from "@/contrib/google/sheets/ui/SheetsFileWidget";
+import SheetsFileWidget from "@/contrib/google/sheets/ui/SpreadsheetPickerWidget";
 import ConnectedFieldTemplate from "@/components/form/ConnectedFieldTemplate";
 import { validateRegistryId } from "@/types/helpers";
 import { type SheetMeta } from "@/contrib/google/sheets/core/types";
